+++ server/controllers/eventsDropdownController.js
